Обновление Laravel с 5.5 до 5.8

+++ database/migrations/2017_09_07_085039_create_places_table.php
@@ -10,7 +10,7 @@ class CreatePlacesTable extends Migration
     public function up()
     {
         Schema::create('places', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name', 40)->nullable()->index()->comment('Имя помещения');
             $table->text('description')->nullable()->comment('Описание для помещения');
@@ -18,7 +18,7 @@ class CreatePlacesTable extends Migration
             $table->bigInteger('phone')->unique()->nullable()->comment('Телефон компании');
             // $table->string('email')->nullable()->comment('Почта');
 
-            $table->integer('location_id')->nullable()->unsigned()->comment('Адрес компании');
+            $table->bigInteger('location_id')->nullable()->unsigned()->comment('Адрес компании');
             $table->foreign('location_id')->references('id')->on('locations');
 
             // $table->integer('schedule_id')->nullable()->unsigned()->comment('Id графика работы');
@@ -30,7 +30,7 @@ class CreatePlacesTable extends Migration
 
 
             // Общие настройки
-            $table->integer('company_id')->unsigned()->nullable()->comment('Id компании');
+            $table->bigInteger('company_id')->unsigned()->nullable()->comment('Id компании');
             $table->foreign('company_id')->references('id')->on('companies');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
@@ -38,7 +38,7 @@ class CreatePlacesTable extends Migration
             $table->integer('system_item')->nullable()->unsigned()->comment('Флаг системной записи: 1 или null');
             $table->integer('moderation')->nullable()->unsigned()->comment('На модерации');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
