+++ database/migrations/2017_11_28_060444_create_category_rights_table.php
@@ -14,14 +14,14 @@ class CreateCategoryRightsTable extends Migration
     public function up()
     {
         Schema::create('category_rights', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('category_right_name')->index()->unique()->comment('Имя категории пользователей');
             
             $table->integer('display')->nullable()->unsigned()->comment('Отображение на сайте');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
