+++ database/migrations/2017_12_07_093006_create_booklists_table.php
@@ -10,23 +10,23 @@ class CreateBooklistsTable extends Migration
     public function up()
     {
         Schema::create('booklists', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name')->nullable()->comment('Имя списка');
             $table->text('description')->nullable()->comment('Описание списка');
 
-            $table->integer('entity_id')->nullable()->unsigned()->comment('Id сущности');
+            $table->bigInteger('entity_id')->nullable()->unsigned()->comment('Id сущности');
             $table->foreign('entity_id')->references('id')->on('entities');
 
             $table->string('entity_alias')->index()->comment('Имя сущности');
 
             $table->boolean('change_allowed')->default(0)->comment('Разрешение на внесение изменений в список с данным типом. Параметр наследуется списку.');
 
-            $table->integer('booklist_type_id')->unsigned()->default(1)->comment('Тип списка');
+            $table->bigInteger('booklist_type_id')->unsigned()->default(1)->comment('Тип списка');
             $table->foreign('booklist_type_id')->references('id')->on('booklist_types');
 
             // Общие настройки
-            $table->integer('company_id')->unsigned()->nullable()->comment('Id компании');
+            $table->bigInteger('company_id')->unsigned()->nullable()->comment('Id компании');
             $table->foreign('company_id')->references('id')->on('companies');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
@@ -34,7 +34,7 @@ class CreateBooklistsTable extends Migration
             $table->integer('system_item')->nullable()->unsigned()->comment('Флаг системной записи: 1 или null');
             $table->integer('moderation')->nullable()->unsigned()->comment('На модерации');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
