+++ database/migrations/2017_12_08_092542_create_booklist_user_table.php
@@ -14,19 +14,19 @@ class CreateBooklistUserTable extends Migration
     public function up()
     {
         Schema::create('booklist_user', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('booklist_id')->nullable()->unsigned()->comment('ID списка');
+            $table->bigInteger('booklist_id')->nullable()->unsigned()->comment('ID списка');
             $table->foreign('booklist_id')->references('id')->on('booklists');
 
-            $table->integer('user_id')->nullable()->unsigned()->comment('ID пользователя');
+            $table->bigInteger('user_id')->nullable()->unsigned()->comment('ID пользователя');
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->integer('display')->nullable()->unsigned()->comment('Отображение на сайте');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
