+++ database/migrations/2017_12_13_044253_create_positions_table.php
@@ -14,19 +14,21 @@ class CreatePositionsTable extends Migration
     public function up()
     {
         Schema::create('positions', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name')->index()->comment('Название должности');
-            $table->integer('page_id')->unsigned()->nullable()->comment('Id страницы приземления');
+
+            $table->bigInteger('page_id')->unsigned()->nullable()->comment('Id страницы приземления');
             $table->foreign('page_id')->references('id')->on('pages');
+            
             $table->integer('direction_status')->unsigned()->nullable()->comment('Руководящая должность или нет');
 
-            $table->integer('sector_id')->nullable()->unsigned()->comment('Id сектора');
+            $table->bigInteger('sector_id')->nullable()->unsigned()->comment('Id сектора');
             $table->foreign('sector_id')->references('id')->on('sectors');
 
 
             // Общие настройки
-            $table->integer('company_id')->unsigned()->nullable()->comment('Id компании');
+            $table->bigInteger('company_id')->unsigned()->nullable()->comment('Id компании');
             $table->foreign('company_id')->references('id')->on('companies');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
@@ -34,7 +36,7 @@ class CreatePositionsTable extends Migration
             $table->integer('system_item')->nullable()->unsigned()->comment('Флаг системной записи: 1 или null');
             $table->integer('moderation')->nullable()->unsigned()->comment('На модерации');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
