+++ database/migrations/2017_12_13_072702_create_staff_table.php
@@ -14,21 +14,22 @@ class CreateStaffTable extends Migration
     public function up()
     {
         Schema::create('staff', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('user_id')->unsigned()->nullable()->comment('Id пользователя');
+            $table->bigInteger('user_id')->unsigned()->nullable()->comment('Id пользователя');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('position_id')->unsigned()->nullable()->comment('Id должности');
+            
+            $table->bigInteger('position_id')->unsigned()->nullable()->comment('Id должности');
             $table->foreign('position_id')->references('id')->on('positions');
 
-            $table->integer('department_id')->unsigned()->nullable()->comment('Id отдела');
+            $table->bigInteger('department_id')->unsigned()->nullable()->comment('Id отдела');
             $table->foreign('department_id')->references('id')->on('departments');
 
-            $table->integer('filial_id')->unsigned()->nullable()->comment('Id отдела');
+            $table->bigInteger('filial_id')->unsigned()->nullable()->comment('Id отдела');
             $table->foreign('filial_id')->references('id')->on('departments');
 
             // Общие настройки
-            $table->integer('company_id')->unsigned()->nullable()->comment('Id компании');
+            $table->bigInteger('company_id')->unsigned()->nullable()->comment('Id компании');
             $table->foreign('company_id')->references('id')->on('companies');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
@@ -36,7 +37,7 @@ class CreateStaffTable extends Migration
             $table->integer('system_item')->nullable()->unsigned()->comment('Флаг системной записи: 1 или null');
             $table->integer('moderation')->nullable()->unsigned()->comment('На модерации');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
