+++ database/migrations/2018_01_26_060314_create_menu_site_table.php
@@ -15,10 +15,10 @@ class CreateMenuSiteTable extends Migration
     {
         Schema::create('menu_site', function (Blueprint $table) {
 
-            $table->integer('menu_id')->nullable()->unsigned()->comment('Id раздела меню');
+            $table->bigInteger('menu_id')->nullable()->unsigned()->comment('Id раздела меню');
             $table->foreign('menu_id')->references('id')->on('menus');
 
-            $table->integer('site_id')->nullable()->unsigned()->comment('Id сайта');
+            $table->bigInteger('site_id')->nullable()->unsigned()->comment('Id сайта');
             $table->foreign('site_id')->references('id')->on('sites');
 
         });
