+++ database/migrations/2018_04_20_061958_create_units_table.php
@@ -14,7 +14,7 @@ class CreateUnitsTable extends Migration
     public function up()
     {
         Schema::create('units', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name')->index()->comment('Название единицы измерения');
             $table->string('abbreviation')->comment('Сокращенное название');
@@ -23,12 +23,12 @@ class CreateUnitsTable extends Migration
 
             $table->text('description')->nullable()->comment('Описание единицы измерения');
 
-            $table->integer('units_category_id')->nullable()->unsigned()->comment('Id категории в которой находиться юнит');
+            $table->bigInteger('units_category_id')->nullable()->unsigned()->comment('Id категории в которой находиться юнит');
             $table->foreign('units_category_id')->references('id')->on('units_categories');
 
 
             // Общие настройки
-            $table->integer('company_id')->unsigned()->nullable()->comment('Id компании');
+            $table->bigInteger('company_id')->unsigned()->nullable()->comment('Id компании');
             $table->foreign('company_id')->references('id')->on('companies');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
@@ -36,7 +36,7 @@ class CreateUnitsTable extends Migration
             $table->integer('system_item')->nullable()->unsigned()->comment('Флаг системной записи: 1 или null');
             $table->integer('moderation')->nullable()->unsigned()->comment('На модерации');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
