+++ database/migrations/2018_04_27_121556_create_processes_groups_table.php
@@ -14,26 +14,26 @@ class CreateProcessesGroupsTable extends Migration
     public function up()
     {
         Schema::create('processes_groups', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name')->index()->comment('Название группы процессов');
             $table->text('description')->nullable()->comment('Описание группы процессов');
 
-            $table->integer('photo_id')->nullable()->unsigned()->comment('Id фото (аватар)');
+            $table->bigInteger('photo_id')->nullable()->unsigned()->comment('Id фото (аватар)');
             $table->foreign('photo_id')->references('id')->on('photos');
 
-            $table->integer('unit_id')->nullable()->unsigned()->comment('Id единицы измерения');
+            $table->bigInteger('unit_id')->nullable()->unsigned()->comment('Id единицы измерения');
             $table->foreign('unit_id')->references('id')->on('units');
 
             $table->integer('rule_id')->nullable()->unsigned()->comment('Id правила определения цены');
             // $table->foreign('rule_id')->references('id')->on('rules');
 
-            $table->integer('album_id')->nullable()->unsigned()->comment('Id альбома');
+            $table->bigInteger('album_id')->nullable()->unsigned()->comment('Id альбома');
             $table->foreign('album_id')->references('id')->on('albums');
 
 
             // Общие настройки
-            $table->integer('company_id')->unsigned()->nullable()->comment('Id компании');
+            $table->bigInteger('company_id')->unsigned()->nullable()->comment('Id компании');
             $table->foreign('company_id')->references('id')->on('companies');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
@@ -41,7 +41,7 @@ class CreateProcessesGroupsTable extends Migration
             $table->integer('system_item')->nullable()->unsigned()->comment('Флаг системной записи: 1 или null');
             $table->integer('moderation')->nullable()->unsigned()->comment('На модерации');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
