+++ database/migrations/2018_04_29_151229_create_workflows_table.php
@@ -14,12 +14,12 @@ class CreateWorkflowsTable extends Migration
     public function up()
     {
         Schema::create('workflows', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('process_id')->nullable()->unsigned()->comment('Id артикула');
+            $table->bigInteger('process_id')->nullable()->unsigned()->comment('Id артикула');
             $table->foreign('process_id')->references('id')->on('processes');
 
-            $table->integer('category_id')->nullable()->unsigned()->comment('Id категории рабочего процесса');
+            $table->bigInteger('category_id')->nullable()->unsigned()->comment('Id категории рабочего процесса');
             $table->foreign('category_id')->references('id')->on('workflows_categories');
 
             $table->boolean('set_status')->default(0)->unsigned()->comment('Статус набора');
@@ -28,7 +28,7 @@ class CreateWorkflowsTable extends Migration
 
 
             // Общие настройки
-            $table->integer('company_id')->unsigned()->nullable()->comment('Id компании');
+            $table->bigInteger('company_id')->unsigned()->nullable()->comment('Id компании');
             $table->foreign('company_id')->references('id')->on('companies');
 
             $table->integer('sort')->nullable()->unsigned()->index()->comment('Поле для сортировки');
@@ -36,7 +36,7 @@ class CreateWorkflowsTable extends Migration
             $table->integer('system_item')->nullable()->unsigned()->comment('Флаг системной записи: 1 или null');
             $table->integer('moderation')->nullable()->unsigned()->comment('На модерации');
 
-            $table->integer('author_id')->nullable()->unsigned()->comment('Id создателя записи');
+            $table->bigInteger('author_id')->nullable()->unsigned()->comment('Id создателя записи');
             $table->foreign('author_id')->references('id')->on('users');
 
             $table->integer('editor_id')->nullable()->unsigned()->comment('Id редактора записи');
